# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Shahebbazar API
  version: 1.1.0-sprint2
  summary: Rajshahi local business directory — Express API.
  description: |
    The Express API behind shahebbazar.com. The Next.js frontend calls it
    from the server when rendering pages; the provider dashboard, the
    analytics beacons and image uploads call it from the browser.

    **Browse it:** run the API and open `http://localhost:4000/api/docs`.
    **Postman:** File → Import → this file (`docs/openapi.yaml`).

    **Authentication is not built yet.** Every provider and admin
    endpoint is marked `TODO(auth)` in the code and is currently open.
    The customer (`/api/me/...`) endpoints act as the first active
    customer account until sign-in exists.
    They must not be deployed publicly until sessions and role checks
    exist. The provider endpoints take the business id in the path for
    now; with sign-in, it will come from the session.

    **Conventions**
    - All public responses contain approved records only.
    - `NUMERIC` and `COUNT(*)` values that PostgreSQL returns as big
      numbers arrive as **strings** (e.g. `"rating": "4.5"`); convert
      before doing arithmetic.
    - Stored image paths are relative (`uploads/…`) and are served by this
      API at `/uploads/…`.
    - Errors are `{ "error": "…" }`; form endpoints add `errors` keyed by
      field.
  contact:
    name: COS40005 Group 23
# No authentication exists yet, so no security scheme is declared. When
...
security: []
servers:
...
paths:
...
  /api/vendors/{vendorId}/listings:
    get:
      operationId: getVendorsByVendorIdListings
      tags: [Listings]
      summary: All of a business's listings, any status
      parameters:
        - $ref: "#/components/parameters/VendorId"
      responses:
        "200":
          description: Newest first.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Listing" } }
    post:
      operationId: postVendorsByVendorIdListings
      tags: [Listings]
      summary: Create a listing (awaits admin approval)
      parameters:
        - $ref: "#/components/parameters/VendorId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ListingInput" }
      responses:
        "201":
          description: Created as pending.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListingSaved" }
        "400": { $ref: "#/components/responses/LegacyError" }
  /api/listings/{listingId}:
    put:
      operationId: putListingsByListingId
      tags: [Listings]
      summary: Update a listing
      parameters:
        - $ref: "#/components/parameters/ListingId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ListingInput" }
      responses:
        "200":
          description: Saved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListingSaved" }
        "400": { $ref: "#/components/responses/LegacyError" }
        "404": { $ref: "#/components/responses/LegacyError" }
    delete:
      operationId: deleteListingsByListingId
      tags: [Listings]
      summary: Delete a listing and its photos
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema: { type: object, properties: { success: { const: true }, message: { type: string } } }
        "404": { $ref: "#/components/responses/LegacyError" }
  /api/listings/{listingId}/photos:
    post:
      operationId: postListingsByListingIdPhotos
      tags: [Listings]
      summary: Upload photos for a listing
      parameters:
        - $ref: "#/components/parameters/ListingId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photos: { type: array, items: { type: string, format: binary, description: "JPG or PNG." } }
      responses:
        "200":
          description: Uploaded.
          content:
            application/json:
              schema: { type: object, properties: { success: { const: true }, message: { type: string } } }
        "400": { $ref: "#/components/responses/LegacyError" }

  # --------------------------------------------------------------- admin
...
  /api/vendors/{vendorId}/owner:
    parameters:
      - $ref: "#/components/parameters/VendorId"
    get:
      operationId: getVendorOwner
      tags: [Provider]
      summary: The owner's name, phone and email
      responses:
        "200": { $ref: "#/components/responses/Owner" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: putVendorOwner
      tags: [Provider]
      summary: Update the owner's name and email (the phone is the sign-in and cannot change here)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_name]
              properties:
                user_name: { type: string, minLength: 2, maxLength: 100 }
                user_email: { type: string, description: Empty removes it. }
      responses:
        "200": { $ref: "#/components/responses/Owner" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: The email is used by another account. }

  # ------------------------------------------------------ admin moderation
...
components:
  parameters:
...
    VendorId:
      name: vendorId
      in: path
      required: true
      description: TODO(auth) — will come from the session.
      schema: { type: integer, minimum: 1 }
      example: 13
    ListingId:
      name: listingId
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
  responses:
...
    Owner:
      description: Owner details.
      content:
        application/json:
          schema:
            type: object
            properties:
              user_id: { type: integer }
              user_name: { type: string }
              user_phone: { type: string }
              user_email: { $ref: "#/components/schemas/NullableString" }
    BadRequest:
      description: Invalid input.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Not found (for public endpoints, also when not approved).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    LegacyError:
      description: Error in the listing/registration routes' format.
      content:
        application/json:
          schema:
            type: object
            required: [success, error]
            properties: { success: { const: false }, error: { type: string } }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        detail: { type: string }
        hint: { type: string }
...
    NumericString:
      type: string
      pattern: '^-?\d+(\.\d+)?$'
      description: A PostgreSQL NUMERIC or BIGINT, as a string.
    NullableString: { type: [string, "null"] }
...
    ListingStatus: { type: string, enum: [draft, pending, approved, rejected, archived] }
...
    Listing:
      type: object
      required: [listing_id, vendor_id, listing_title, listing_slug, listing_status]
      properties:
        listing_id: { type: integer }
        vendor_id: { type: integer }
        category_id: { type: [integer, "null"] }
        listing_title: { type: string }
        listing_title_bn: { $ref: "#/components/schemas/NullableString" }
        listing_slug: { type: string }
        listing_description: { $ref: "#/components/schemas/NullableString" }
        listing_price: { oneOf: [{ $ref: "#/components/schemas/NumericString" }, { type: "null" }] }
        listing_price_max: { oneOf: [{ $ref: "#/components/schemas/NumericString" }, { type: "null" }] }
        listing_price_unit: { $ref: "#/components/schemas/NullableString" }
        listing_min_order_qty: { type: [integer, "null"] }
        listing_is_negotiable: { type: boolean }
        listing_status: { $ref: "#/components/schemas/ListingStatus" }
        listing_rejection_reason: { $ref: "#/components/schemas/NullableString" }
        listing_created_at: { type: string, format: date-time }
        category_name: { $ref: "#/components/schemas/NullableString" }
        category_slug: { $ref: "#/components/schemas/NullableString" }
    ListingInput:
      type: object
      required: [listing_title, category_id]
      properties:
        listing_title: { type: string }
        listing_title_bn: { type: string }
        listing_description: { type: string }
        category_id: { type: integer }
        listing_price: { type: [number, "null"], description: "Leave out for \"price on request\"." }
        listing_price_max: { type: [number, "null"] }
        listing_price_unit: { type: string }
        listing_min_order_qty: { type: integer }
        listing_is_negotiable: { type: boolean, default: true }
    ListingSaved:
      type: object
      required: [success, listing_id, listing_slug, listing_status]
      properties:
        success: { const: true }
        listing_id: { type: integer }
        listing_slug: { type: string }
        listing_status: { $ref: "#/components/schemas/ListingStatus" }
...